+++ src/components/pages/diary/diary/Content.tsx
@@ -1,5 +1,5 @@
 import EmotionTag from "../../../common/EmotionTag";
-import HeartIcon from "../../../assets/svg/heart.svg?react";
+import HeartIcon from "@/assets/svg/heart.svg?react";
 import { useEffect, useRef, useState } from "react";
 import { addLike, removeLike } from "../../../../api/api";
 import { useParams } from "react-router-dom";
